requirements.txt
pyspark==3.5.0
pandas==2.1.0
psycopg2-binary==2.9.9
apache-airflow==2.8.0
pyyaml==6.0.1
pyarrow==14.0.0
